==> code/2_logistic_regression/5_categorical_data_logistic_regression.ipynb <==
import pandas as pd

data_url= "http://www-stat.wharton.upenn.edu/~waterman/DataSets/uva.txt"
dataframe = pd.read_table(data_url)

dataframe[:5]

dataframe.shape

dataframe.dtypes

dataframe["Marital Status"].unique()

dataframe.pop('who') # dataframe = dataframe.ix[:, 1:]
dataframe.pop('Country')
dataframe.pop('Years on Internet')

dataframe.dtypes

dataframe["Gender"] = dataframe["Gender"].astype("category")
dataframe.dtypes

for col in ['Household Income', 'Sexual Preference', 'Education Attainment', 'Major Occupation', "Marital Status"]:
    dataframe[col] = dataframe[col].astype('category')
dataframe.dtypes    

dataframe_modified = pd.get_dummies(dataframe)
dataframe_modified[:5]

dataframe_modified.shape

dataframe_modified['Age'][pd.isnull(dataframe_modified['Age'])] = dataframe_modified['Age'].mean()

x_data = dataframe_modified.ix[:, 1:].values
y_data = dataframe_modified.ix[:, 0].reshape(-1,1)

y_data.shape, x_data.shape

from sklearn import preprocessing # Min-Max Standardzation

min_max_scaler = preprocessing.MinMaxScaler()
x_data[:, 0] = min_max_scaler.fit_transform(x_data[:, 0])

x_data[:5,0]

import numpy as np 

training_idx = np.random.randint(y_data.shape[0], size=int(y_data.shape[0] * 0.8))
test_idx = np.random.randint(y_data.shape[0], size=int(y_data.shape[0] * 0.2))

x_training, x_test = x_data[training_idx,:], x_data[test_idx,:]
y_training, y_test = y_data[training_idx,:], y_data[test_idx,:]

x_training.shape, x_test.shape

from sklearn import linear_model, datasets

logreg = linear_model.LogisticRegression(fit_intercept=False)
logreg.fit(x_training, y_training.ravel())

sum(logreg.predict(x_test) == y_test.ravel())  / y_test.shape[0]

training_idx = np.random.randint(y_data.shape[0], size=int(y_data.shape[0] * 0.8))
test_idx = np.random.randint(y_data.shape[0], size=int(y_data.shape[0] * 0.2))

x_training, x_test = x_data[training_idx,:], x_data[test_idx,:]
y_training, y_test = y_data[training_idx,:], y_data[test_idx,:]

x_training.shape, x_test.shape

import tensorflow as tf

X = tf.placeholder(tf.float32)
Y = tf.placeholder(tf.float32)

W = tf.Variable(tf.random_uniform([len(x_data[0]), 1], -1.0, -1.0))

h = tf.matmul(X, W)
hypothesis = tf.div(1., 1. + tf.exp(-h))
cost = -1 * tf.reduce_mean(Y * tf.log(hypothesis) + (1-Y) * tf.log(1-hypothesis))

a = tf.Variable(0.1)
optimizer = tf.train.GradientDescentOptimizer(a)
train = optimizer.minimize(cost)

init = tf.initialize_all_variables()

sess = tf.Session()
sess.run(init)


cost_history = []

for i in range(20000):
    sess.run(train, feed_dict={X: x_training, Y:y_training})
    if i % 2000 == 0:
        print (i, sess.run(cost, feed_dict={X: x_training, Y:y_training}))
        cost_history.append(sess.run(cost, feed_dict={X: x_training, Y:y_training})) 
final_weight = sess.run(W)         
sess.close()

def sigmoid(z):
    return 1 / (1 + np.exp(z))

def hypothesis_function(x, theta):
    z = (np.dot(-x,theta))    
    return sigmoid(z)

sum((hypothesis_function(x_test, final_weight) > 0.5) == y_test) / y_test.shape[0]
