==> code/2_logistic_regression/6_multinomail_logistic_regression.ipynb <==
import pandas as pd
import numpy as np

data_url = "http://www.ats.ucla.edu/stat/data/hsbdemo.dta"
df = pd.read_stata(data_url)
df[:5] 

df["awards"].unique()

df.pop("id")
df.pop("cid")
df[:5]

df.dtypes

y_data = np.reshape(df.pop("prog").values, (-1.1))
y_data.shape

x_data = pd.get_dummies(df).values
x_data[:5]

from sklearn import preprocessing # Min-Max Standardzation

min_max_scaler = preprocessing.MinMaxScaler()
x_data[:, :6] = min_max_scaler.fit_transform(x_data[:, :6])

x_data[:5,:6]

import numpy as np 

training_idx = np.random.randint(y_data.shape[0], size=int(y_data.shape[0] * 0.8))
test_idx = np.random.randint(y_data.shape[0], size=int(y_data.shape[0] * 0.2))

x_training, x_test = x_data[training_idx,:], x_data[test_idx,:]
y_training, y_test = y_data[training_idx], y_data[test_idx]

x_training.shape, x_test.shape

from sklearn import linear_model, datasets

logreg = linear_model.LogisticRegression(multi_class='multinomial', fit_intercept=True, solver="lbfgs")
logreg.fit(x_training, y_training)

sum(logreg.predict(x_test) == y_test)  / y_test.shape[0] 

data_url = "http://www.ats.ucla.edu/stat/data/hsbdemo.dta"
df = pd.read_stata(data_url)
df.pop("id")
df.pop("cid")
df[:5]

y_data = pd.get_dummies(df.ix[:,"prog"]).as_matrix()
y_data[:5,:]

x_data_df = df.drop(["prog"],axis=1)
x_data_df.insert(0, "bias", 1)
x_data_df.ix[:5,:]

x_data = pd.get_dummies(x_data_df).as_matrix()
x_data[:5]

from sklearn import preprocessing # Min-Max Standardzation

min_max_scaler = preprocessing.MinMaxScaler()
x_data[:, 1:7] = min_max_scaler.fit_transform(x_data[:, 1:7])

x_data[:5,:7]

training_idx = np.random.randint(y_data.shape[0], size=int(y_data.shape[0] * 0.8))
test_idx = np.random.randint(y_data.shape[0], size=int(y_data.shape[0] * 0.2))

x_training, x_test = x_data[training_idx,:], x_data[test_idx,:]
y_training, y_test = y_data[training_idx,:], y_data[test_idx,:]

x_training.shape, x_test.shape

import tensorflow as tf

X = tf.placeholder("float", [None, len(x_data[0])])
Y = tf.placeholder("float", [None, len(y_data[0])])

W = tf.Variable(tf.zeros([len(x_data[0]), len(y_data[0])]))

hypothesis = tf.nn.softmax(tf.matmul(X,W))
learning_rate = 0.001

cost = tf.reduce_mean(-1 * tf.reduce_sum(Y * tf.log(hypothesis), reduction_indices=1))
optimizer = tf.train.GradientDescentOptimizer(learning_rate).minimize(cost)

init = tf.initialize_all_variables()

sess = tf.Session()
sess.run(init)
    
for i in range(400001):
    sess.run(optimizer, feed_dict={X: x_training, Y:y_training})
    if i % 40000 == 0:
        print ("%d 's iteration" % i)
        print (sess.run(cost, feed_dict={X: x_training, Y:y_training}))
finale_theta = sess.run(W)


hypotehsis_value = sess.run(hypothesis, feed_dict={X:x_test})
result= [np.argmax(predict) == np.argmax(original_value)  for predict, original_value in zip(hypotehsis_value, y_test)]
sum(result) / len(result)
